Rename file
InstallRuleInstaller class was recently renamed to
InstallRulePluginInstaller to match the conventions in the codebase.
Rename file to match the exported class.
