+++ src/installers/InstallRulePluginInstaller.ts
